requirements.txt
fastapi
uvicorn
python-jose
sentence-transformers
faiss-cpu
numpy
